Add new import helpers
